pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.7.4",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langsmith>=0.4.15",
    "nest-asyncio>=1.6.0",
]
